change the type of img_array into float64

+++ num_gradient/num_gradient_descent_master/t_ngd_cifar10.py
@@ -294,8 +294,7 @@ def create_f(h, w, target):
 def linearize_pixels(img):
     x = np.copy(np.asarray(img))
     h, w, c = x.shape
-    #img_array = x.reshape(h * w * c).astype('float64')
-    img_array = x.reshape(h * w * c).astype(np.uint8)
+    img_array = x.reshape(h * w * c).astype('float64')
     return h, w, img_array
